+++ set_version.py
@@ -1,0 +1,33 @@
+#!/usr/bin/env python3
+"""
+Sets the version in both pyproject.toml and version.py to the same version.
+
+Usage:
+    ./set_version.py <version>
+"""
+import sys
+import os
+import os.path as osp
+
+REPO_DIR = osp.dirname(osp.abspath(__file__))
+LIB_DIR = osp.join(REPO_DIR, 'open2fa')
+
+
+def main(v: str):
+    """sets both pyproject.toml and version.py to the same version"""
+    with open(osp.join(LIB_DIR, 'version.py'), 'w') as f:
+        f.write('__version__ = "{}"\n'.format(v))
+        print('Set version in version.py to', v)
+    with open(osp.join(REPO_DIR, 'pyproject.toml'), 'r') as f:
+        lines = f.readlines()
+        for i, line in enumerate(lines):
+            if line.startswith('version = "'):
+                lines[i] = 'version = "{}"\n'.format(v)
+                break
+    with open(osp.join(REPO_DIR, 'pyproject.toml'), 'w') as f:
+        f.writelines(lines)
+    print('Set version in pyproject.toml to', v)
+
+
+if __name__ == '__main__':
+    main(sys.argv[1])